+++ examples/setup-temp/islands/Counter.tsx
@@ -1,5 +1,5 @@
 import type { Signal } from "@preact/signals";
-import { s } from "../../../mod.ts";
+import { Button } from "../../../mod.ts";
 import { handleInteraction } from "../../../src/utils.ts";
 
 interface CounterProps {
@@ -9,19 +9,17 @@ interface CounterProps {
 export default function Counter(props: CounterProps) {
   return (
     <div class="flex gap-8 py-one-and-half items-center">
-      <button
-        class={s.btn.panel}
+      <Button.Panel
         {...handleInteraction(() => props.count.value -= 1)}
       >
         -1
-      </button>
-      <span class={`tabular-nums ${s.txt.subhead}`}>{props.count}</span>
-      <button
-        class={s.btn.panel}
+      </Button.Panel>
+      <span class="tabular-nums text-subhead">{props.count}</span>
+      <Button.Panel
         {...handleInteraction(() => props.count.value += 1)}
       >
         +1
-      </button>
+      </Button.Panel>
     </div>
   );
 }