+++ mod.ts
@@ -1,7 +1,10 @@
-export * from './src/styles.ts';
-import * as styles from './src/styles.ts';
+export * from './src/atoms.tsx';
 
-export const s = styles.s;
+export { default as InputFieldCombo } from './components/InputFieldCombo.tsx';
+export { default as InputRadioCombo } from './components/InputRadioCombo.tsx';
+export { default as InputCheckCombo } from './components/InputCheckCombo.tsx';
+export { default as InputTextareaCombo } from './components/InputTextareaCombo.tsx';
+export { default as InputSelectCombo } from './components/InputSelectCombo.tsx';
 
 // import renderInput, {
 //   type iInput,